refactor/minimize top layer
azure_diagnostic_logs.js & azure_activity_log.js only differed in 2
lines of code. extract as much common code as possible.
